Inline InvokeMethod within CaseExtensions, now that it is only used once.
